% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_extract_MS2.R
\name{batch_extract_MS2}
\alias{batch_extract_MS2}
\title{Extract the most MS/MS intense scan from batch spectra}
\usage{
batch_extract_MS2(batch_spect, verbose = TRUE, out_list = TRUE)
}
\arguments{
\item{batch_spect}{a list created by \code{\link[=batch_import_mzxml]{batch_import_mzxml()}}}

\item{verbose}{a boolean indicating if the MS/MS TIC chromatogram is displayed}

\item{out_list}{a boolean expressing if the output is a list containing
the MS/MS spectra plus the TIC chromatogram (out_list = TRUE), or only
the data frame with the MS/MS spectra (out_lists = FALSE).}
}
\description{
This function takes multiple spectra imported with the \code{\link[=batch_import_mzxml]{batch_import_mzxml()}}
function. Then, it passes the spectra to a version of \code{\link[=extract_MS2]{extract_MS2()}}.
Briefly, this function takes a series of MS/MS spectra, selects the
most intense scan and extracts the MS/MS spectra from it.
Additionally, it plots the MS/MS TIC chromatogram and colors
the most intense scan with a red circle, and the precursor ion with a blue
diamond
}
\examples{

# Select the csv file name and path
batch_file <- system.file("extdata", "batch_read.csv",
  package = "MS2extract"
)
# Read the data frame
batch_data <- read.csv(batch_file)

# File paths for Procyanidin A2 and Rutin
ProcA2_file <- system.file("extdata",
  "ProcyanidinA2_neg_20eV.mzXML",
  package = "MS2extract"
)
Rutin_file <- system.file("extdata",
  "Rutin_neg_20eV.mzXML",
  package = "MS2extract"
)

# Add file path - User should specify the file path -
batch_data$File <- c(ProcA2_file, Rutin_file)

# Checking batch_data data frame
batch_data

# Using batch import to import multiple compounds
batch_compounds <- batch_import_mzxml(batch_data)

# Checking dimension by compound
# Procyanidin A2: 24249 ions
# Rutin: 22096 ions
purrr::map(batch_compounds, dim)

# Use extract batch extract_MS2
batch_extracted <- batch_extract_MS2(batch_compounds,
  verbose = TRUE,
  out_list = FALSE
)
}
