% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CE_check.R
\name{CE_filter}
\alias{CE_filter}
\title{Filter Collision Energy Values}
\usage{
CE_filter(spec, spec_metadata)
}
\arguments{
\item{spec}{MS/MS spectra}

\item{spec_metadata}{MS/MS spectra metadata containing \code{COLLISIONENERGY} as
one of the provided columns.}
}
\description{
\itemize{
\item Internal Function *
}
}
\details{
This functions evaluate the collision energy value provided in the
\code{spe_metadata} data frame, and matches this value with the collision
energy value found in the MS/MS data. This function uses
\code{\link[readr:parse_number]{readr::parse_number()}}  to extract the numeric value provided in
the \code{COLLISIONENERGY} column.
}
