% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mzxml.R
\name{check_metadata}
\alias{check_metadata}
\title{Check compound metadata}
\usage{
check_metadata(met_metadata)
}
\arguments{
\item{met_metadata}{a data frame with at least the Formula and the
Ionization_mode}
}
\value{
a double with the ionized mass of the compound.
}
\description{
This function evaluates the information provided by the user whether the
right column were provided as well the right ionization modes. Then
it calculates the neutral exact mass of the compound by the given
formula and depending of the ionization mode, it adds or substract the mass
of a proton to obtain the ionized mass of the compounds.
}
