% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_checks.R
\name{check_MS_data_order}
\alias{check_MS_data_order}
\title{Check if spec MS/MS data and spec_metadata is proper aligned}
\usage{
check_MS_data_order(spec = NULL, spec_metadata = NULL)
}
\arguments{
\item{spec}{a data frame containing the extracted MS/MS spectra,
the following columns are required:

\describe{
\item{mz_precursor}{}
\item{rt}{}
\item{mz}{}
\item{intensity}{}
}}

\item{spec_metadata}{a data frame containing the values to be including
in the resulting \code{.msp} file. The following columns are required as vital
information for the .msp output.

\describe{
\item{NAME}{}
\item{PRECURSORTYPE}{}
\item{FORMULA}{}
\item{RETENTIONTIME}{}
\item{IONMODE}{}
\item{COLLISIONENERGY}{}
}}
}
\description{
\itemize{
\item Internal Function *
}
}
\details{
This function is intended to check the order of the key of the spec
data and the spec metadata. If both keys are aligned, the code continues.
If the keys are not aligned, it aborts. This function will not try to align
the keys, as the the keys are ordered back-end and key alignment is expected.

Checks in this function include:
\itemize{
\item Check equal number of MS/MS data and number of observations in spec_metadata
\item All key names are aligned
}
}
