% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_checks.R
\name{check_specdt_mgf_gnps}
\alias{check_specdt_mgf_gnps}
\title{Check spec_metadata for GNPS \code{.mgf} format before exporting the MS/MS library}
\usage{
check_specdt_mgf_gnps(spec_metadata = NULL)
}
\description{
\emph{Internal function}
}
\details{
This functions aims to create a unique KEY to sort and iterate the compound
table in order to match the keys created when MS/MS was imported in
\code{sort_compound_table()}.

The three main steps are:
\itemize{
\item Create a unique row key: Compound name + Polarity + CE
\item Check if there are any duplicated values
\item Sort the compound table by Name and CE value
}
}
