% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_MS2.R
\name{extract_MS2}
\alias{extract_MS2}
\title{Extract the most MS2 intense scan}
\usage{
extract_MS2(spec, verbose = TRUE, out_list = TRUE, ...)
}
\arguments{
\item{spec}{a data frame with the MS2 spectra}

\item{verbose}{a boolean indicating if the MS2 TIC chromatogram is displayed}

\item{out_list}{a boolean expressing if the output is a list containing
the MS2 spectra plus the TIC chromatogram (verbose = TRUE), or only
the data frame with the MS2 spectra (verbose = FALSE).}

\item{...}{any other arguments passed to detect_mass() function}
}
\description{
This function takes raw MS2 signal and looks for the most intense scan
and extracts the MS2 spectra. Additionally, it plots the MS2 TIC
chromatogram and color with red the most intense scan.
}
\examples{
# Importing the Spectrum of Procyanidin A2 in negative ionzation mode
# and 20 eV as the collision energy
ProcA2_file <- system.file("extdata",
                       "ProcyanidinA2_neg_20eV.mzXML",
                        package = "MS2extract")

# Region of interest table (rt in seconds)
ProcA2_data <- data.frame(Formula = "C30H24O12",Ionization_mode = "Negative",
                     min_rt = 163, max_rt = 180)
# Importing MS2 data
ProcA2_raw <- import_mzxml(ProcA2_file, ProcA2_data)

# Extracting most intense scan ----
# Returning plot + MS2 spectra
extract_MS2(ProcA2_raw)

# Returning MS2 spectra only
extract_MS2(ProcA2_raw, out_list = FALSE)
}
