% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mzxml.R
\name{import_mzxml}
\alias{import_mzxml}
\title{Imports mzxml files with MS2 scans}
\usage{
import_mzxml(file, roi_table = NULL, ...)
}
\arguments{
\item{file}{file name of the mzxml file}

\item{roi_table}{a data frame with two columns min_rt and max_rt specifying
the minimum and maximum retention time range (in seconds).}

\item{...}{extra arguments passed to  masstools::read_mzxml()}
}
\value{
data.frame in a tidy format for MS2 spectra in a tidy format.
 \describe{
  \item{mz}{ion m/z value}
  \item{intensity}{ion intensity count}
  \item{precursor_mz}{precursor ion}
  \item{rt}{retention time (in seconds)}
}
}
\description{
This function imports MS2 data using the masstools::read_mzxml()
function to import the data as a list. Each element in a list *list* represents
one scan. Subsequently, each element in this list (scans) contains two
sublists that gathers the scan information and the spectra per scan.
}
\details{
This function will import all the scans contained in the provided mzxml
file, and users can use filters on the resulting data to extract the
desired information.
}
\examples{

\donttest{
# Importing the Spectrum of Procyanidin A2 in negative ionization mode
# and 20 eV as the collision energy
ProcA2_file <- system.file("extdata",
                       "ProcyanidinA2_neg_20eV.mzXML",
                        package = "MS2extract")

ProcA2_raw <- import_mzxml(ProcA2_file)

# 26731 ions detected in total
dim(ProcA2_raw)

# Region of interest table (rt in seconds)
ROI_dt <- data.frame(min_rt = 163, max_rt = 180)
ProcA2_roi <- import_mzxml(ProcA2_file, roi_table = ROI_dt)

# 24249 ions detected in ROI
dim(ProcA2_roi)

}
}
