% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MS2spectra.R
\name{plot_MS2spectra}
\alias{plot_MS2spectra}
\title{Plot MS2 spectra}
\usage{
plot_MS2spectra(spec, compound = NULL, ppm = 10)
}
\arguments{
\item{spec}{a data frame containing the MS2 spectra of the most intense
scan}

\item{compound}{a character, if user is using batch_* functions, user need
to provide a character with the given name of the compound stated in the
batch_import table.}

\item{ppm}{ppm error tolerance to check if the mm/z precursor is being
detected or not in the MS2 spectra.}
}
\value{
a ggplot plot. Filled blue diamond means that the precursor ion
was detected in the MS2 spectra. If the precursor ion was not detected in
the MS2 spectra, the diamond is not filled.
}
\description{
This functions plot the resulting MS2 spectra of the most intense scan
}
