% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_MS2.R
\name{plot_tic}
\alias{plot_tic}
\title{Plot MS2 TIC}
\usage{
plot_tic(TIC)
}
\arguments{
\item{TIC}{a data frame with the TIC}
}
\value{
a list. The first element is a data frame with the retention time
with the highest intensity and the TIC plot as the second list element.
}
\description{
This function plots the MS2 TIC and will mark the most intense scan
that will be used to extract the MS2 spectra
}
