% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ppm.R
\name{ppm_range}
\alias{ppm_range}
\title{Calculate the \emph{m}/\emph{z} upper and lower values limits given a ppm tolerance}
\usage{
ppm_range(mz, ppm)
}
\arguments{
\item{mz}{a double, the theoretical \emph{m}/\emph{z} value.}

\item{ppm}{a double, the mass error in ppm}
}
\value{
a vector of two elements containing the upper and lower \emph{m}/\emph{z}values
given the provided mass error tolerance.
}
\description{
Given a theoretical \emph{m}/\emph{z} value, and a ppm error tolerance,
this function calculates the lower and upper \emph{m}/\emph{z} values.
}
\examples{

chlorogenic_acid_pos <- 355.1023
ppm_error <- 10
ppm_range(mz = chlorogenic_acid_pos, ppm = ppm_error)
}
