% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi_filter.R
\name{roi_filter}
\alias{roi_filter}
\title{Filtering the region of interest (ROI)}
\usage{
roi_filter(spec, roi_table)
}
\arguments{
\item{spec}{a data frame containing}

\item{roi_table}{a data frame with two columns min_rt and max_rt specifying
the minimum and maximum retention time for a specific metabolite.}
}
\description{
This functions takes a data frame with the minimum and maximum retention time
(in seconds), and keeps the \emph{scans} inside the provided boundaries. This
filter functions aims to keep the scans between the provided ROI and remove
the scans outside the ROI.
}
\examples{
\donttest{
# Importing the Spectrum of Procyanidin A2 in negative ionization mode
# and 20 eV as the collision energy
ProcA2_file <- system.file("extdata",
                       "ProcyanidinA2_neg_20eV.mzXML",
                        package = "MS2extract")

# Region of interest table (rt ins seconds)
ProcA2_data <- data.frame(Formula = "C30H24O12",Ionization_mode = "Negative",
                     min_rt = 163, max_rt = 180)

ProcA2_raw <- import_mzxml(ProcA2_file, ProcA2_data)

# 26731 ions detected in total
dim(ProcA2_raw)
}
}
