% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_checks.R
\name{sort_compound_table}
\alias{sort_compound_table}
\title{Sort compound table before importing data}
\usage{
sort_compound_table(compounds_dt = NULL)
}
\arguments{
\item{compounds_dt}{a data frame containing the following columns.
\describe{
\item{Name}{The name of the compound}
\item{Formula}{The compound's chemical formula}
\item{Ionization_mode}{The ionization mode set in data collection
(only Positive and Negative modes allowed).}
\item{File}{The filename of the \code{.mzXML}/\code{.mzML} file including the path}
\item{COLLISIONENERGY}{Collision energy applied in MS/MS fragmentation}
}}
}
\description{
\emph{Internal function}
}
\details{
This function checks for the compound data to have unique values
(metabolites) per row, before importing the MS/MS data.
This function performs the following checks:
\itemize{
\item Create a unique row key: Compound name + Polarity + CE
\item Check if there are any duplicated values based on the keys
\item Sort the compound table by Name and CE value
}
}
