% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mgf.R
\name{write_gnps_table}
\alias{write_gnps_table}
\title{Creates the batch upload table for GNPS batch upload}
\usage{
write_gnps_table(spec, spec_metadata, mgf_filename)
}
\arguments{
\item{spec}{extracted spectra}

\item{spec_metadata}{spectra metadata}

\item{mgf_filename}{file name of the .mgf library. This name is created in
the calling function.}
}
\description{
\itemize{
\item Internal Function *
}
}
\details{
This function is designed to create the batch upload table required by GNPS
\href{https://ccms-ucsd.github.io/GNPSDocumentation/batchupload}{batch upload}
job. This function takes both, \code{spec} data and the \code{spec_metadata} to create
the bath upload table.
}
