% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mgf.R
\name{write_mgf_gnps}
\alias{write_mgf_gnps}
\title{Create the mgf backbone file format}
\usage{
write_mgf_gnps(spec = NULL, spec_metadata = NULL, mgf_name = NULL)
}
\arguments{
\item{spec}{a data frame containing the extracted MS2 spectra, the following
colums are required:

\describe{
\item{mz_precursor}{}
\item{rt}{}
\item{mz}{}
\item{intensity}{}
}}

\item{spec_metadata}{a data frame containing the values to be including
in the resulting mgf file. In this case, this is the minimum mandatory
information to be included.

\describe{
\item{INCHIKEY}{}
\item{SMILES}{}
\item{CCS}{}
\item{COLLISIONENERGY}{}
\item{INSTRUMENTTYPE}{}
}}
}
\description{
\emph{Internal function}
}
\details{
This functions takes an extracted MS2 spectra and creates the backbone
of a typical mgf file based on
\href{http://www.matrixscience.com/help/data_file_help.html}{MATRIXSCIENCE}
information.
}
