% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_msp.R
\name{write_msp}
\alias{write_msp}
\title{Export MS2 spectra to a msp file}
\usage{
write_msp(spec = NULL, spec_metadata = NULL, msp_name = NULL)
}
\arguments{
\item{spec}{a data frame containing the extracted MS2 spectra, the following
colums are required:

\describe{
 \item{mz_precursor}{}
 \item{rt}{}
 \item{mz}{}
 \item{intensity}{}
}}

\item{spec_metadata}{a data frame containing the values to be including
in the resulting msp file. The following column are required as vital
information for the msp output.

\describe{
 \item{NAME}{}
 \item{PRECURSORTYPE}{}
 \item{FORMULA}{}
 \item{RETENTIONTIME}{}
 \item{IONMODE}{}
}

The follosing fields are also included in the resulting msp files, but are
not requiered to be present in the metadata table. If the column does not
exist in the column or the value is missing, it will export a blank field.
 \describe{
  \item{INCHIKEY}{}
  \item{SMILES}{}
  \item{CCS}{}
  \item{COLLISIONENERGY}{}
  \item{INSTRUMENTTYPE}{}
}}

\item{msp_name}{a string with the name of the msp file not containing (.msp)}
}
\description{
Thins functions takes an extracted MS2 spectra and writes in a msp file
format. This fuctions takes the extracted MS2 spectra and the metadata
for the compound. It will draw information such as the precursor m/z
as well as the retention time of the scan from the spec data frame and
the rest of information from the provided metadata.
}
\examples{
\dontrun{
# Importing the Spectrum of Procyanidin A2 in negative ionization mode
# and 20 eV as the collision energy
ProcA2_file <- system.file("extdata",
                       "ProcyanidinA2_neg_20eV.mzXML",
                        package = "MS2extract")

 # Region of interest table (rt in seconds)
ProcA2_data <- data.frame(Formula = "C30H24O12",Ionization_mode = "Negative",
                     min_rt = 163, max_rt = 180)

ProcA2_raw <- import_mzxml(ProcA2_file, ProcA2_data)

# Detecting masses with the normalized spectra and ions with
# intensities greater than 5\%
ProcA2_detected <- detect_mass(ProcA2_raw, normalize = TRUE, min_int = 1)

# Reading the metadata
metadata_file <- system.file("extdata",
                             "msp_metadata.csv",
                              package = "MS2extract")
metadata <- read.csv(metadata_file)

# Exporting msp file
write_msp(spec = ProcA2_detected,
          spec_metadata = metadata,
          msp_name = "Procyanidin_A2")
}
}
