% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_msp.R
\name{add_attributes}
\alias{add_attributes}
\title{Evaluate additional msp fields}
\usage{
add_attributes(msp_attribute, spec_metadata, msp_backbone)
}
\arguments{
\item{msp_attribute}{a string with the name of the msp field}

\item{spec_metadata}{a data frame containing the msp field values}

\item{msp_backbone}{the constructed msp string to export}
}
\description{
This internal functions evaluate non-critical msp fields to be included in the
msp file. This function will create an empty field if any of the
evaluated fields are either NA value in the metadata table or the
column is missing.
}
\details{
The evaluated additional msp fields are:

\describe{
\item{INCHIKEY}{}
\item{SMILES}{}
\item{CCS}{}
\item{COLLISIONENERGY}{}
\item{INSTRUMENTTYPE}{}
}
}
