% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mzxml.R
\name{assign_scan_id}
\alias{assign_scan_id}
\title{Extract MS2 spectrum info out of the list}
\usage{
assign_scan_id(scan_list)
}
\arguments{
\item{scan_list}{a list created by masstools::read_mzxml()}
}
\value{
a data.frame containing
\describe{
\item{mz}{ion \emph{m}/\emph{z} value}
\item{intensity}{ion intensity count}
\item{precursor_mz}{precursor ion}
\item{rt}{retention time}
}
}
\description{
This function extracts spectra per scan in a tidy format given a list
created by masstools::read_mzxml()
}
