% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_extract_MS2.R
\name{batch_extract_MS2}
\alias{batch_extract_MS2}
\title{Extract the most MS2 intense scan from batch spectra}
\usage{
batch_extract_MS2(batch_spect, verbose = TRUE, out_list = TRUE)
}
\arguments{
\item{batch_spect}{a list created by batch_import_mzxml()}

\item{verbose}{a Boolean indicating if the MS2 TIC chromatogram is displayed}

\item{out_list}{a Boolean expressing if the output is a list containing
the MS2 spectra plus the TIC chromatogram (verbose = TRUE), or only
the data frame with the MS2 spectra (verbose = FALSE).}
}
\description{
This function takes multiple spectra imported with batch_import_mzxml()
function. Then, it passes the argument the argument to extract_MS2()
function. Briefly, this function takes raw MS2 signal and looks
for the most intense scan and extracts the MS2 spectra.
Additionally, it plots the MS2 TIC chromatogram and color
with red the most intense scan.
}
\examples{

# Select the csv file name and path
batch_file <- system.file("extdata", "batch_read.csv",
                          package = "MS2extract")
# Read the data frame
batch_data <- read.csv(batch_file)

# File paths for Procyanidin A2 and Rutin
ProcA2_file <- system.file("extdata",
                          "ProcyanidinA2_neg_20eV.mzXML",
                           package = "MS2extract")
Rutin_file <- system.file("extdata",
                         "Rutin_neg_20eV.mzXML",
                          package = "MS2extract")

# Add file path - User should specified the file path -
batch_data$File <- c(ProcA2_file, Rutin_file)

# Checking batch_data data frame
batch_data

# Using batch import to import multiple compounds
batch_compounds <- batch_import_mzxml(batch_data)

# Checking dimension by compound
# Procyanidin A2: 24249 ions
# Rutin: 22096 ions
purrr::map(batch_compounds, dim)

# Use extract batch extract_MS2
batch_extracted <- batch_extract_MS2(batch_compounds,
                                     verbose = TRUE,
                                     out_list = FALSE)
}
