% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mzxml.R
\name{check_metadata}
\alias{check_metadata}
\title{Check compound metadata}
\usage{
check_metadata(met_metadata)
}
\arguments{
\item{met_metadata}{a data frame with at least the Formula and the
Ionization_mode}
}
\value{
a double with the ionized mass of the compound.
}
\description{
\emph{Internal function}
}
\details{
This function evaluates if the metadata has the expected column names.
Then, it calculates
the theoretical exact mass of the compound using the given formula and
ionization mode to obtain a charged mass.

Required values

\describe{
\item{Formula}{character, compound chemical formula}
\item{Ionization_mode}{character,
only \emph{Positive} and \emph{Negative} values are acepted }
}
}
