% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mzxml.R
\name{check_metadata}
\alias{check_metadata}
\title{Check compound metadata}
\usage{
check_metadata(met_metadata)
}
\arguments{
\item{met_metadata}{a data frame with at least the Formula and the
Ionization_mode}
}
\value{
a double with the ionized mass of the compound.
}
\description{
This function evaluates the information provided by the user whether the
correct column  and ionization modes were provided. Then, it calculates
the neutral exact mass of the compound using the given formula and
ionization mode to obtain a charged mass.
}
