% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_checks.R
\name{check_specdt_msp}
\alias{check_specdt_msp}
\title{Check spec_metadata for NIST format before exporting the MS/MS library}
\usage{
check_specdt_msp(spec_metadata = NULL)
}
\arguments{
\item{spec_metadata}{a data frame containing the values to be including
in the resulting \code{.mgf} file. In this case, this is the minimum mandatory
information to be included.}
}
\description{
\emph{Internal function}
}
\details{
This function aims to create a unique KEY to sort and iterate the compound
table in order to match the keys created when MS/MS was imported in
\code{sort_compound_table()}.

The three main steps are:
\itemize{
\item Create a unique row key: Compound name + Polarity + CE
\item Check if there are any duplicated values
\item Sort the compound table by Name and CE value
}
}
