% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mzxml.R
\name{extract_scan_info}
\alias{extract_scan_info}
\title{Extract scan info out of the list}
\usage{
extract_scan_info(mzxml)
}
\arguments{
\item{mzxml}{a list created by masstools::read_mzxml()}
}
\value{
a data.frame containing
\describe{
\item{name}{the scan name (m/z and rt)}
\item{mz}{the precursor m/z ion}
\item{rt}{retention time}
\item{file name}{file name}
}
}
\description{
Given a list created by masstools::read_mzxml(), this function
extract the scan info for all scans in a tidy format.
}
