% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ppm.R
\name{get_ppm}
\alias{get_ppm}
\title{Calculate the  \emph{m}/\emph{z} difference for a given mass and ppm tolerance}
\usage{
get_ppm(mz, ppm = 10)
}
\arguments{
\item{mz}{a double, the theoretical m/z value.}

\item{ppm}{a double, the ppm error tolerance.}
}
\value{
a double with the m/z value deviation from the given
theoretical m/z value
}
\description{
Given a theoretical  \emph{m}/\emph{z} value and ppm tolerance,
it computes the  \emph{m}/\emph{z} deviation.
}
\examples{
get_ppm(mz = 355.1023, ppm = 20)

# Chlorogenic acid [M+H]+ = 355.1023 m/z
chlorogenic_acid_pos <- 355.1023
ppm_error <- 10
get_ppm(mz = chlorogenic_acid_pos, ppm = ppm_error)
}
