% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_MS2.R
\name{get_TIC}
\alias{get_TIC}
\title{Calculate the TIC}
\usage{
get_TIC(spec)
}
\arguments{
\item{spec}{a data fram with two columns, mz and intensity}
}
\description{
This function compute the MS2 TIC of a spectra grouping by retention time.
}
