% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mzxml.R
\name{import_mzxml}
\alias{import_mzxml}
\title{Imports mzXML/mzML files with MS2 scans}
\usage{
import_mzxml(file = NULL, met_metadata = NULL, ppm = 10, ...)
}
\arguments{
\item{file}{file name and path of the .mzXML or .mzML MS/MS data}

\item{met_metadata}{a data frame with the following columns.

\strong{Required}:
\describe{
\item{Formula}{A character string specifying the metabolite formula}
\item{Ionization_mode}{The ionization mode employed in data collection. It
can be only Positive or Negative}
}

These two columns are mandatory since the formula is used by the Rdisop
package to calculate the exact mass and the ionization mode
determine whether the the mass of the a proton is added or subtracted to
calculate the charged mass.

Additionally, you can provide the minimum and maximum retention times
to look for a peak only within a given area by including the following
columns:

\describe{
\item{min_rt}{numeric, with the minimum retention time to keep}
\item{max_rt}{numeric, with the minimum retention time to keep}

}

These two columns are highly recommended to be included to narrow down
the search window and ensure the peak you want is selected. This is
especially important if you have multiple peaks in the same data file.}

\item{ppm}{the mass error in ppm. 10 ppm is the default value.}

\item{...}{extra arguments passed to  MS2extract:::read_mzxml()}
}
\value{
data.frame in a tidy format for MS2 spectra in a tidy format.
\describe{
\item{mz}{ion m/z value}
\item{intensity}{ion intensity count}
\item{mz_precursor}{precursor ion}
\item{rt}{retention time (in seconds)}
\item{Formula}{Chemical formula provided in \code{met_metada}}
}
}
\description{
This function reads .mzXML and .mzML files containing MS2. This function is
inspired on \code{masstools::read_mzxml()}which imports the data as a list. Then,
each element in a list list represents one scan. Then, each scans contains
two sub-lists that contain (1) the scan information and
(2) the spectra per scan.
}
\examples{
\donttest{
# Importing the Spectrum of Procyanidin A2 in negative ionization mode
# and 20 eV as the collision energy
ProcA2_file <- system.file("extdata",
  "ProcyanidinA2_neg_20eV.mzXML",
  package = "MS2extract"
)
# Compound metadata without ROI information
ProcA2_data <- data.frame(Formula = "C30H24O12", Ionization_mode = "Negative")
ProcA2_raw <- import_mzxml(ProcA2_file, met_metadata = ProcA2_data)

# 26731 ions detected in total
dim(ProcA2_raw)

# Region of interest table (rt in seconds)
ProcA2_data <- data.frame(
  Formula = "C30H24O12", Ionization_mode = "Negative",
  min_rt = 163, max_rt = 180
)
ProcA2_roi <- import_mzxml(ProcA2_file, met_metadata = ProcA2_data)

# 24249 ions detected in ROI
dim(ProcA2_roi)
}
}
