% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2extract-package.R
\docType{package}
\name{MS2extract-package}
\alias{MS2extract}
\alias{MS2extract-package}
\title{MS2extract: Targeted MS2 spectra extraction}
\description{
This package has the objective to import mass spectrometry data and extract the desired MS2 spectrum data and perform basics filters such as remove low intensity counts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cooperstonelab.github.io/MS2extract/}
}

}
\author{
\strong{Maintainer}: Cristian Quiroz-Moreno \email{cristianquirozd1997@gmail.com} (\href{https://orcid.org/0000-0002-9069-9147}{ORCID})

}
\keyword{internal}
