% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2extract-package.R
\docType{package}
\name{MS2extract-package}
\alias{MS2extract}
\alias{MS2extract-package}
\title{MS2extract: Targeted MS2 Spectra Extraction for In-house Compound Library Creation}
\description{
A mass spectrometry toolbox with the objective of importing mass spectrometry data, and extracts the desired MS2 spectra for in-house MS/MS compound library creation. Subsequently, users can export the spectra in a .msp file format, as a resulting compound database for further use in compound identification tasks. It is worth to note that this package do not annotate mass spectrometry data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cooperstonelab.github.io/MS2extract/}
  \item Report bugs at \url{https://github.com/CooperstoneLab/MS2extract/issues}
}

}
\author{
\strong{Maintainer}: Cristian Quiroz-Moreno \email{cristianquirozd1997@gmail.com} (\href{https://orcid.org/0000-0002-9069-9147}{ORCID})

Authors:
\itemize{
  \item Jessica Cooperstone \email{cooperstone.1@osu.edu} (\href{https://orcid.org/0000-0001-7920-0088}{ORCID})
}

}
\keyword{internal}
