% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_spec.R
\name{normalize_spec}
\alias{normalize_spec}
\title{Normalization and low intensity signal filtering}
\usage{
normalize_spec(spec, min_int = 1)
}
\arguments{
\item{spec}{a data frame produced by import_mzxml() or with the same columns}

\item{min_int}{a integer the minimum normalized ion intensity between
1 and 100%.}
}
\value{
a data frame with the normalized ion intensities and signal greater
than the provided minimum intensity.
}
\description{
This function takes a raw spectra (ion abundance in total counts) and
normalized by the base peak. Additionally, this function filters the
signal by a minimum intensity to remove background noise. By default,
the minimum intensity is 1%.
}
\examples{

# Importing the Spectrum of Procyanidin A2 in negative ionzation mode
# and 20 eV as the collision energy
ProcA2_file <- system.file("extdata",
                       "ProcyanidinA2_neg_20eV.mzXML",
                        package = "MS2extract")

ProcA2_raw <- import_mzxml(ProcA2_file)

# Check raw intensities
range(ProcA2_raw$intensity) # [1]  1.125 851439.500

normalize_spec
}
