% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_spec.R
\name{normalize_spec}
\alias{normalize_spec}
\title{Normalization and low-intensity signal filtering}
\usage{
normalize_spec(spec, min_int = 1)
}
\arguments{
\item{spec}{a data frame produced by \code{import_mzxml()} function.}

\item{min_int}{an integer, minimum intensity. If spectra are normalized,
this value should be in a 0-100 range. If the spectra are
not normalized, this value must be in raw ion count.}
}
\description{
This function normalizes raw spectra by converting from raw intensities,
to 0-100\% relative intensity based on the most intense ion.
This function also removes low-intensity signals based on the minimum cutoff.
The default cutoff is 1\% and the spectra are normalized.
}
\examples{
# Importing the Spectrum of Procyanidin A2 in negative ionization mode
# and 20 eV as the collision energy
ProcA2_file <- system.file("extdata",
  "ProcyanidinA2_neg_20eV.mzXML",
  package = "MS2extract"
)

# Region of interest table (rt in seconds)
ProcA2_data <- data.frame(
  Formula = "C30H24O12", Ionization_mode = "Negative",
  min_rt = 163, max_rt = 180
)

ProcA2_raw <- import_mzxml(ProcA2_file, ProcA2_data)

# Check raw intensities
# Ranges:  1.125 851439.500
range(ProcA2_raw$intensity)

ProcA2_norm <- normalize_spec(ProcA2_raw)
# Ranges: 2 100
range(ProcA2_norm$intensity)
}
