% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MS2spectra.R
\name{plot_MS2base}
\alias{plot_MS2base}
\title{Base function for MS2 spectra}
\usage{
plot_MS2base(spec, ppm)
}
\arguments{
\item{spec}{a data frame containing the MS2 spectra of the most intense
scan}

\item{ppm}{ppm error tolerance to check if the mm/z precursor is being
detected or not in the MS2 spectra}
}
\description{
\emph{Internal function}
}
\details{
This is the base function for plotting the MS2 spectra.
}
