% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_MS2.R
\name{plot_MS2spectra}
\alias{plot_MS2spectra}
\title{Plot MS2 spectra}
\usage{
plot_MS2spectra(spec)
}
\arguments{
\item{spec}{a data frame containing the MS2 spectra of the most intense
scan}
}
\value{
a ggplot plot
}
\description{
This functions plot the resulting MS2 spectra of the most intense scan
}
