% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MS2spectra.R
\name{plot_MS2spectra}
\alias{plot_MS2spectra}
\title{Plot MS2 spectra}
\usage{
plot_MS2spectra(spec, compound = NULL, ppm = 10)
}
\arguments{
\item{spec}{a data frame containing the MS2 spectra of the most intense
scan}

\item{compound}{a character, if user is using batch_* functions, they need
to provide a character with the identical “Name” of the compound provided
in the \code{batch_import_mzxml()} table.}

\item{ppm}{mass error in ppm tolerance to check if the mm/z precursor is being
detected or not in the MS2 spectra.}
}
\value{
a \code{ggplot} plot of the MS2 spectra. A filled blue diamond is
placed above the precursor ion. If the precursor ion was not
detected in the MS2 spectra, the blue diamond is not filled.
}
\description{
This functions plots the resulting MS2 spectra of the most intense scan
}
