% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MS2spectra.R
\name{plot_MS2spectra}
\alias{plot_MS2spectra}
\title{Plot MS/ms spectra}
\usage{
plot_MS2spectra(spec, compound = NULL, ppm = 10)
}
\arguments{
\item{spec}{a data frame containing the MS/MS spectra of the most intense
scan}

\item{compound}{a character, if user is using batch_* functions, they need
to provide a character with the identical “Name” of the compound provided
in the \code{batch_import_mzxml()} table.}

\item{ppm}{mass error in ppm tolerance to check if the \emph{m}/\emph{z} precursor is being
detected or not in the MS/MS spectra.}
}
\value{
a \code{ggplot} plot of the MS/MS spectra. A filled blue diamond is
placed above the precursor ion. If the precursor ion was not
detected in the MS/MS spectra, the blue diamond is not filled.
}
\description{
This function plots the resulting MS/MS spectra of the most intense scan
}
\examples{

Rutin_file <- system.file("extdata",
  "Rutin_neg_20eV.mzXML",
   package = "MS2extract"
 )

 # Region of interest table (rt in seconds)
Rutin_data <- data.frame(Formula = "C27H30O16",
    Ionization_mode = "Negative",
    min_rt = 160, max_rt = 175
 )
 # Importing MS2 data
rutin_raw <- import_mzxml(Rutin_file, Rutin_data)
Rutin_extracted <- extract_MS2(rutin_raw)

Rutin_detected <- detect_mass(Rutin_extracted,
   normalize = TRUE, # Allow normalization
   min_int = 1) # 1\% as minimum intensity
MS2_spectra <- plot_MS2spectra(Rutin_detected)
print(MS2_spectra)
}
