% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_MS2.R
\name{plot_tic}
\alias{plot_tic}
\title{Plot MS2 TIC}
\usage{
plot_tic(spec)
}
\arguments{
\item{spec}{a data frame with the MS/MS spectra}
}
\value{
a list. The first element is a data frame with the retention time
with the highest intensity and the EIC plot as the second list element.
}
\description{
\emph{Internal function}
}
\details{
This function plots the MS2 TIC and will mark the most intense scan
that will be used to extract the MS/MS spectra
}
