% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ppm.R
\name{ppm_range}
\alias{ppm_range}
\title{Calculate the m/z upper and lower values limits given a ppm tolerance}
\usage{
ppm_range(mz, ppm)
}
\arguments{
\item{mz}{a double, the theoretical m/z value.}

\item{ppm}{a double, the ppm error tolerance}
}
\value{
a vector of two elements containing the upper and lower
m/z values with the specific ppm error tolerance.
}
\description{
Given a theoretical m/z value and a ppm error tolerance,
this function calculates the lower and upper m/z values.
}
\examples{
# Chlorogenic acid [M+H]+ = 355.1023 m/z
chlorogenic_acid_pos <- 355.1023
ppm_error = 10
ppm_range(mz = chlorogenic_acid_pos, ppm = ppm_error)
}
