% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mzxml.R
\name{read_mzxml}
\alias{read_mzxml}
\title{Read mzxml/mzML files containing MS2 data}
\usage{
read_mzxml(file, threads = 3, mode = c("inMemory", "onDisk"))
}
\arguments{
\item{file}{The vector of names of ms2 files. MS2 file must be mzXML or mzML.}

\item{threads}{Thread number}

\item{mode}{inMemory or onDisk}
}
\value{
Return ms2 data. This is a list.
}
\description{
This function takes a mzxml file and imports the MS2 data only
to further create a list contaning to sublists with the scan information
(fisrt sublits) and the MS2 spectra (second sublist)
}
\examples{
ProcA2_file <- system.file("extdata",
                       "ProcyanidinA2_neg_20eV.mzXML",
                        package = "MS2extract")

ProcA2_raw <- read_mzxml(ProcA2_file)
}
\author{
Xiaotao Shen inspired from the masstools package
\email{shenxt1990@outlook.com}
}
