% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mgf.R
\name{write_mgf_gnps}
\alias{write_mgf_gnps}
\title{Create the GNPS mgf backbone file format}
\usage{
write_mgf_gnps(spec = NULL, spec_metadata = NULL, mgf_name = NULL)
}
\arguments{
\item{spec}{a data frame containing the extracted MS2 spectra, the following
colums are required:

\describe{
\item{mz_precursor}{precursor ion}
\item{rt}{retention time}
\item{mz}{\emph{m/z}  values}
\item{intensity}{intensity values}
}}

\item{spec_metadata}{a data frame containing the values to be including
in the resulting mgf file. In this case, this is the minimum mandatory
information to be included.

The full explanation about fields and field content can be found in
GNPS batch library upload
\href{https://ccms-ucsd.github.io/GNPSDocumentation/batchupload}{link}.

For the rest of fields that are included in the final library file,
MS2extract will get this information for the extracted spectra.

We highly suggest to check the \code{gnps_template.xlsx} file described in
the function example to check what fields/columns are required in
MS2extract.

\describe{
\item{COMPOUND_NAME}{character, Metabolite name, it has to be the same
name used in \code{met_metadata}, the data frame used to import your MS/MS data}
\item{INSTRUMENT}{character, instrument used for data collection}
\item{COLLISIONENERGY}{character, collision energy used in MS/MS
fragmentation}
\item{IONSOURCE}{Ionization source}
\item{SMILES}{character, SMILES chemical structure of your metabolite}
\item{INCHI}{character, Inchi value for the metabolite}
\item{INCHIAUX}{character, INCHIAUX for the metabolite}
\item{IONMODE}{character, ionization polarity}
\item{PUBMED}{character, PUBMED id where you submitted the MS/MS spectra}
\item{ACQUISITION}{character, Crude, Lysate, Commercial, Isolated, Other}
\item{DATACOLLECTOR}{character, Person who collected the MS/MS data}
\item{INTEREST}{character, interest of the MS/MS data}
\item{LIBQUALITY}{character, library quality.
1 for Gold, 2 for Silver, 3 for Bronze}
\item{GENUS}{character, genus of the organism}
\item{SPECIES}{character, species of the organism}
\item{STRAIN}{character, strain of the organism}
\item{CASNUMBER}{character, CAS number of the metabolite}
\item{PI}{character, principal investigator}
}}

\item{mgf_name}{file name for the exported \code{mgf}library. It does not have
to contain the file extension  \code{.mgf}.}
}
\value{
if batch spectra are found, this function writes two files,
the \code{.mgf} library and the required \code{.tsv} table required by GNPS. If single
spectrum is detected, it will only write the \code{.mgf} library.
}
\description{
\emph{Internal function}
}
\details{
This function facilitates to create the structure of the GNPS .mgf format.
For more information about submitting your spectra to GNPS, please
visit this \href{https://ccms-ucsd.github.io/GNPSDocumentation/batchupload}{link}.

You can find the GNPS template spreadsheet in this
\href{https://ccms-ucsd.github.io/GNPSDocumentation/static/Template.xlsx}{link}.
}
\examples{
# Example with batch spectra ----


# Select the csv file name and path
batch_file <- system.file("extdata", "batch_read.csv",
  package = "MS2extract"
)
# Read the data frame
batch_data <- read.csv(batch_file)

# File paths for Procyanidin A2 and Rutin
ProcA2_file <- system.file("extdata",
  "ProcyanidinA2_neg_20eV.mzXML",
  package = "MS2extract"
)
Rutin_file <- system.file("extdata",
  "Rutin_neg_20eV.mzXML",
  package = "MS2extract"
)

# Add file path - User should specified the file path -
batch_data$File <- c(ProcA2_file, Rutin_file)

# Checking batch_data data frame
batch_data

# Using batch import to import multiple compounds
batch_compounds <- batch_import_mzxml(batch_data)
# Checking dimension by compound
# Procyanidin A2: 24249 ions
# Rutin: 22096 ions
purrr::map(batch_compounds, dim)

batch_extracted_compounds <- batch_extract_MS2(batch_compounds)

# Batch detect mass
batch_mass_detected <- batch_detect_mass(batch_extracted_compounds,
  normalize = TRUE, # Normalize
  min_int = 1 # Minimum intensity
)

# Reading metadata from GNPS template
template_file <- system.file("extdata", "GNPS_template.xlsx",
                 package = "MS2extract")

gnps_template <- readxl::read_excel(path = template_file,
                sheet = "batch_example")

write_mgf_gnps(spec = batch_mass_detected,
               spec_metadata = gnps_template,
               mgf_name = "PhenolicsDB")

}
