% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_msp.R
\name{write_msp}
\alias{write_msp}
\title{Export MS2 spectra to a .msp file}
\usage{
write_msp(spec = NULL, spec_metadata = NULL, msp_name = NULL)
}
\arguments{
\item{spec}{a data frame containing the extracted MS2 spectra, the following
colums are required:

\describe{
\item{mz_precursor}{}
\item{rt}{}
\item{mz}{}
\item{intensity}{}
}}

\item{spec_metadata}{a data frame containing the values to be including
in the resulting .msp file. The following columns are required as vital
information for the .msp output.

\describe{
\item{NAME}{}
\item{PRECURSORTYPE}{}
\item{FORMULA}{}
\item{RETENTIONTIME}{}
\item{IONMODE}{}
}

The following fields can be included in the resulting .msp file, but are
not required to be present in the metadata table. If the column does not
exist in the column or the value is missing, a blank field will be exported
\describe{
\item{INCHIKEY}{}
\item{SMILES}{}
\item{CCS}{}
\item{COLLISIONENERGY}{}
\item{INSTRUMENTTYPE}{}
}}

\item{msp_name}{a string with the name of the .msp file excluding the file extension}
}
\description{
This functions takes an extracted MS2 spectra and writes it to a .msp
file format. This function incorporates the extracted MS2 spectra along
with metadata for the compound.
}
\examples{

# Example with single MS2 spectra -----

# Importing the Spectrum of Procyanidin A2 in negative ionization mode
# and 20 eV as the collision energy
ProcA2_file <- system.file("extdata",
  "ProcyanidinA2_neg_20eV.mzXML",
  package = "MS2extract"
)

# Region of interest table (rt in seconds)
ProcA2_data <- data.frame(
  Formula = "C30H24O12", Ionization_mode = "Negative",
  min_rt = 163, max_rt = 180
)

ProcA2_raw <- import_mzxml(ProcA2_file, ProcA2_data)

# Extracting the most intense MS2 scan
ProcA2_ext <- extract_MS2(ProcA2_raw)

# Detecting masses with the normalized spectra and ions with
# intensities greater than 1\%
ProcA2_detected <- detect_mass(ProcA2_ext,
  normalize = TRUE, # Allow normalization
  min_int = 1
) # 1\% as minimum intensity

# Reading the metadata
metadata_file <- system.file("extdata",
  "msp_metadata.csv",
  package = "MS2extract"
)
metadata <- read.csv(metadata_file)

# Exporting msp file
write_msp(
  spec = ProcA2_detected,
  spec_metadata = metadata,
  msp_name = "Procyanidin_A2"
)



# Example with batch spectra ----


# Select the csv file name and path
batch_file <- system.file("extdata", "batch_read.csv",
  package = "MS2extract"
)
# Read the data frame
batch_data <- read.csv(batch_file)

# File paths for Procyanidin A2 and Rutin
ProcA2_file <- system.file("extdata",
  "ProcyanidinA2_neg_20eV.mzXML",
  package = "MS2extract"
)
Rutin_file <- system.file("extdata",
  "Rutin_neg_20eV.mzXML",
  package = "MS2extract"
)

# Add file path - User should specified the file path -
batch_data$File <- c(ProcA2_file, Rutin_file)

# Checking batch_data data frame
batch_data

# Reading batch metadata
metadata_msp_file <- system.file("extdata",
  "batch_msp_metadata.csv",
  package = "MS2extract"
)

metadata_msp <- read.csv(metadata_msp_file)

# Using batch import to import multiple compounds
batch_compounds <- batch_import_mzxml(batch_data)
# Checking dimension by compound
# Procyanidin A2: 24249 ions
# Rutin: 22096 ions
purrr::map(batch_compounds, dim)

batch_extracted_compounds <- batch_extract_MS2(batch_compounds)

# Batch detect mass
batch_mass_detected <- batch_detect_mass(batch_extracted_compounds, # Compound list
  normalize = TRUE, # Normalize
  min_int = 1
) # Minimum intensity

# Checking dimension by compound
# Procyanidin A2: 107 ions
# Rutin: 12 ions
purrr::map(batch_mass_detected, dim)

# Writing msp file
write_msp(
  spec = batch_mass_detected,
  spec_metadata = metadata_msp,
  msp_name = "ProcA2_Rutin_batch"
)
}
